added recux flow for <OrderBook />, reformatted all files properly
