+++ src/components/OrderBook/OrderBookRenderer.js
@@ -1,46 +1,47 @@
 // @flow
 import React from 'react';
-import { OrderBook } from './OrderBook';
-import styled from 'styled-components';
-import { sortArray } from '../../utils/helpers';
-type Props = {
-  loading: boolean,
-  decimals: number,
-  sellOrderList: Array<Object>,
-  buyOrderList: Array<Object>,
-  quoteToken: string,
-  baseToken: string,
+import Loading from '../Loading';
+import { reduceDecimals } from '../../utils/converters';
+import type { OrderListTypes, SingleOrderTypes } from '../../types/orderBook';
+
+type ListTypes = OrderListTypes;
+type SingleOrderProps = SingleOrderTypes;
+
+export const OrderBookRenderer = (props: ListTypes) => {
+  const { bookName, loading, quoteToken, baseToken, decimals, orderList } = props;
+  return (
+    <div className={bookName + ' order-book inner pt-dark'}>
+      <h5>{bookName}</h5>
+      {loading && <Loading />}
+      {!loading && (
+        <div className="list-container">
+          <ul className="pt-list-unstyled heading">
+            <li className="heading">
+              <span className="index">#</span>
+              <span className="total">Total ({quoteToken})</span>
+              <span className="amount">Amount ({baseToken})</span>
+              <span className="price">Price ({quoteToken})</span>
+            </li>
+          </ul>
+          <ul className="pt-list-unstyled list">
+            {orderList.map((order, index) => (
+              <SingleOrder decimals={decimals} key={index} index={index} order={order} />
+            ))}
+          </ul>
+        </div>
+      )}
+    </div>
+  );
 };
 
-export default class OrderBookRenderer extends React.PureComponent<Props> {
-  static defaultProps = {
-    decimals: 2,
-  };
-  render() {
-    const { buyOrderList, sellOrderList, loading, baseToken, quoteToken, decimals } = this.props;
-    return (
-      <Row>
-        <OrderBook
-          orderList={sortArray(sellOrderList, 'price')}
-          bookName="Sell"
-          loading={loading}
-          baseToken={baseToken}
-          quoteToken={quoteToken}
-          decimals={decimals}
-        />
-        <OrderBook
-          orderList={sortArray(buyOrderList, 'price')}
-          bookName="Buy"
-          loading={loading}
-          baseToken={baseToken}
-          quoteToken={quoteToken}
-          decimals={decimals}
-        />
-      </Row>
-    );
-  }
-}
-const Row = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
+const SingleOrder = (props: SingleOrderProps) => {
+  const { order, index, decimals } = props;
+  return (
+    <li className="not-heading">
+      <span className="index">{index + 1}</span>
+      <span className="total">{reduceDecimals(order.amount * order.price, decimals)}</span>
+      <span className="amount">{reduceDecimals(order.amount, decimals)}</span>
+      <span className="price">{reduceDecimals(order.price, decimals)}</span>
+    </li>
+  );
+};
