+++ src/store/createReducer.js
@@ -1,5 +1,5 @@
 // @flow
-import type { State, Action, ActionHandler } from './index';
+import type { Action, ActionHandler, State } from './index';
 
 export default function createReducer(actionHandler: ActionHandler) {
   return (state: State, action: Action) => actionHandler(action)(state);
