+++ src/store/index.js
@@ -1,4 +1,4 @@
-import { applyMiddleware, createStore as baseCreateStore, combineReducers, compose } from 'redux';
+import { applyMiddleware, combineReducers, compose, createStore as baseCreateStore } from 'redux';
 import thunk from 'redux-thunk';
 import * as reducers from './reducers';
 import '../styles/css/index.css';
