+++ src/store/models/createWallet.test.js
@@ -1,8 +1,8 @@
 import { createStore } from '../../store';
 
 import * as actionCreators from './createWallet';
-import * as walletService from '../services/wallet';
 import getCreateWalletModel from './createWallet';
+import * as walletService from '../services/wallet';
 
 jest.mock('../services/wallet');
