+++ src/store/models/homePage.js
@@ -4,14 +4,28 @@ import type { LoadDataParams } from '../../types/homePage';
 import type { State, ThunkAction } from '../../types';
 import { getData } from '../services/homePage';
 import * as ohlcvActionCreators from '../actions/ohlcv';
+import * as orderBookActionCreators from '../actions/orderBook';
+
+import * as orderList from '../../jsons/ordersList.json';
 
 export default function getHomePageModel(state: State) {
   return HomePageModel(state.homePage);
 }
 
+const orderBookData = {
+  buyOrderList: orderList.list,
+  sellOrderList: orderList.list,
+  baseToken: 'ETH',
+  quoteToken: 'USDT',
+  loading: false,
+  decimals: 7,
+};
+
 export const loadData = ({ tokenId }: LoadDataParams): ThunkAction => {
   return async (dispatch, getState) => {
     let ohlcvData = await getData();
     dispatch(ohlcvActionCreators.saveData(ohlcvData));
+
+    dispatch(orderBookActionCreators.saveData(orderBookData));
   };
 };
