+++ src/store/services/homePage.js
@@ -1,4 +1,4 @@
-import { tsvParse, csvParse } from 'd3-dsv';
+import { tsvParse } from 'd3-dsv';
 import { timeParse } from 'd3-time-format';
 
 export function parseData(parse) {
