+++ src/types/homePage.js
@@ -1,16 +1,6 @@
 //@flow
 
-export type HomeState = {
-  +loading: boolean,
-  +decimals: number,
-  +quoteToken: string,
-  +baseToken: string,
-  +sellOrderList: Array<Object>,
-  +buyOrderList: Array<Object>,
-  +ohlcvData: Array<Object>,
-  +orderHistory: Array<Object>,
-  +tradeHistory: Array<Object>,
-};
+export type HomeState = {};
 
 export type LoadDataParams = {
   tokenId: string,
