+++ src/types/index.js
@@ -5,6 +5,7 @@ import type { EtherBalanceState, EtherBalanceAction, EtherBalanceEvent } from '.
 import type { AccountBalancesState, AccountBalancesEvent, AccountBalancesAction } from './accountBalances';
 import type { EtherTxState, EtherTxAction, EtherTxEvent } from './etherTx';
 import type { OHLCVState } from './ohlcv';
+import type { OrderBookState } from './orderBook';
 import type { HomeState } from './homePage';
 import type { TokenState, TokenAction, TokenEvent } from './tokens';
 import type { AccountState, AccountAction, AccountEvent } from './account';
@@ -46,6 +47,7 @@ export type State = {
   etherTx: EtherTxState,
   homePage: HomeState,
   ohlcv: OHLCVState,
+  orderBook: OrderBookState,
   etherTx: EtherTxState,
   tokens: TokenState,
   account: AccountState,
