+++ src/types/orderBook.js
@@ -1,5 +1,13 @@
 //@flow
 
+export type OrderBookState = {
+  +loading: boolean,
+  +decimals: number,
+  +sellOrderList: Array<Object>,
+  +buyOrderList: Array<Object>,
+  +quoteToken: string,
+  +baseToken: string,
+};
 export type OrderListTypes = {
   orderList: Array<Object>,
   bookName: string,
